pixman_back/requirements.txt
django
django-rest-framework==0.1.0
django-cors-headers
psycopg2
Pillow